% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chl_eur11.R
\name{chl_eur11}
\alias{chl_eur11}
\title{Access High resolution climate data for Europe}
\usage{
chl_eur11(period = "year", var, year, path = ".")
}
\arguments{
\item{period}{time period (see below).}

\item{var}{variable name (see XXX).}

\item{year}{year only available when period is set to 'daily' or 'yearly'.}

\item{path}{path to the folder where files will be stored.}
}
\description{
Access High resolution climate data for Europe
}
\details{
Time periods (\code{periods}) are as follows:
\itemize{
\item \code{daily}: Data files on a daily resolution (usually years available are 1981-2005)
\item \code{yearly}: Annual aggregations of a variable (usually years available are 1981-2005).
\item \code{monthly}: Data files on a monthly resolution.
\item \code{normal}: Long term, climatological, means of a variable over a Normals period.
}
}
\examples{
\dontrun{
chl_eur11("normal", "pr")
}
}
\references{
\itemize{
\item \url{https://os.zhdk.cloud.switch.ch/envicloud/chelsa/chelsa_V2/EUR11/documentation/CHELSA_EUR11_technical_documentation.pdf}
\item Karger, D.N., Conrad, O., Böhner, J., Kawohl, T., Kreft, H., Soria-Auza, R.W., Zimmermann, N.E., Linder, H.P. & Kessler, M. (2017) Climatologies at high resolution for the earth’s land surface areas. Scientific Data 4, 170122
\item Karger, Dirk Nikolaus; Dabaghchian, Babek; Lange, Stefan; Thuiller, Wilfried; Zimmermann, Niklaus E.; Graham, Catherine H. (2020). High resolution climate data for Europe. EnviDat. doi:10.16904/envidat.150.
}
}
