% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chl_global.R
\name{chl_global_mod}
\alias{chl_global_mod}
\alias{chl_global_obsv_daily}
\alias{chl_global_obsv_monthly}
\title{Chelsa V2}
\usage{
chl_global_mod(
  horizon = 2040,
  var = "bio",
  val = 10,
  model = "mpi-esm1",
  scenario = "ssp126",
  path = "."
)

chl_global_obsv_daily(var, month = 1, day = 1, path = ".")

chl_global_obsv_monthly(var, month = 1, year = 1980, path = ".")
}
\arguments{
\item{horizon}{time horizons.}

\item{var}{variable.}

\item{val}{values.}

\item{model}{models.}

\item{scenario}{scenarios.}

\item{path}{path to the folder where files will be stored.}

\item{month}{month.}

\item{day}{day.}

\item{year}{year.}
}
\description{
Chelsa V2
}
\details{
\itemize{
\item Models:
\itemize{
\item \code{mpi-esm1} : MPI-ESM1-2-HR
\item \code{gfdl-esm4} : GFDL-ESM4
\item \code{ipsl-cm6a} : IPSL-CM6A-LR
\item \code{mri-esm2} : MRI-ESM2-0
\item \code{ukesm1} : UKESM1-0-LL
}
\item Scenarios (\code{scenarios}): SSP (Shared Socioeconomic Pathways)
\itemize{
\item \code{ssp126}: SSP 1, 2.6 W.m-2
\item \code{ssp370}: SSP 3, 7.0 W.m-2
\item \code{ssp585}: SSP 5, 8.5 W.m-2
}
\item Time horizons (\code{horizons}):
\itemize{
\item \code{2010}: 1981-2010 ('model' and 'scenario' are ignored).
\item \code{2040}: 2011-2040
\item \code{2070}: 2041-2070
\item \code{2100}: 2071-2100
}
}
}
\section{Functions}{
\itemize{
\item \code{chl_global_obsv_daily()}: observed data daily (for 1980).

\item \code{chl_global_obsv_monthly()}: observed data monthly (from 1980 up to 2018) .

}}
\examples{
\dontrun{
 chl_global_mod(var = "pr", val = 1, model = "ukesm1", scenario = "ssp370", horizon = 2010) 
 chl_global_mod(var = "pr", val = 1, model = "ukesm1", scenario = "ssp370", horizon = 2040) 
 chl_global_obsv_monthly(var = "pr")
}
}
