% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chl_ts.R
\name{chl_ts}
\alias{chl_ts}
\title{Access CHELSA timeseries}
\usage{
chl_ts(var, year, val = NULL, path = ".")
}
\arguments{
\item{var}{variable name.}

\item{year}{year.}

\item{val}{optional.}

\item{path}{path to the folder where files will be stored.}
}
\description{
Access CHELSA timeseries
}
\details{
"The CHELSA timeseries data consists of monthly downscaled model output temperature and precipitation estimates at a horizontal resolution of 30 arc sec. from 1979-2013. The resulting data consist of a mean monthly temperature and precipitation amounts." (Karger et al. 2017)
}
\examples{
\dontrun{
chl_ts("prec", 1982, 3)
chl_ts("gts30", 1982)
}
}
\references{
\itemize{
\item \url{https://chelsa-climate.org}
\item Karger, D.N., Conrad, O., Böhner, J., Kawohl, T., Kreft, H., Soria-Auza, R.W., Zimmermann, N.E., Linder, H.P., Kessler, M.(2017) Data from: Climatologies at high resolution for the earth’s land surface areas. Dryad Digital Repository. doi.org/10.5061/dryad.kd1d4.
}
}
