% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chl_v1.R
\name{get_chelsea_data}
\alias{get_chelsea_data}
\title{Download Chelsea data}
\usage{
get_chelsea_data(categ = "clim", type = "bio", id = 1, path = ".")
}
\arguments{
\item{categ}{category.}

\item{type}{type.}

\item{id}{resource id.}

\item{path}{path to the folder where files will be stored.}
}
\description{
Download Chelsea data
}
\references{
\url{https://chelsa-climate.org}
}
